--- a/.gitignore
+++ b/.gitignore
@@ -130,3 +130,5 @@ dmypy.json
 
 # vscode
 .vscode/
+
+**/*DS_Store*