--- a/notebooks/measure_time.ipynb
+++ b/notebooks/measure_time.ipynb
@@ -1,0 +1,140 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Measure time taken to construct the singling out attacks"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 13,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import os\n",
+    "import pandas as pd\n",
+    "import matplotlib.pyplot as plt\n",
+    "import seaborn as sns\n",
+    "\n",
+    "from anonymeter.evaluators import SinglingOutEvaluator\n",
+    "from anonymeter.evaluators.singling_out_evaluator import _evaluate_queries, _generate_singling_out_queries"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Data"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "dataset_dir = \"./datasets/\"\n",
+    "\n",
+    "ori = pd.read_csv(os.path.join(dataset_dir, \"adults_train.csv\"))\n",
+    "syn = pd.read_csv(os.path.join(dataset_dir, \"adults_syn_ctgan.csv\"))\n",
+    "control = pd.read_csv(os.path.join(dataset_dir, \"adults_control.csv\"))"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Construct the evaluator"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "max_attempts = 1_000_000\n",
+    "evaluator = SinglingOutEvaluator(ori=ori, \n",
+    "                                 syn=syn, \n",
+    "                                 control=control,\n",
+    "                                 n_attacks=max_attempts,\n",
+    "                                 max_attempts=max_attempts)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Run only the part of the evaluation that is concerned with constructing & running queries"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "mode = \"multivariate\""
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 12,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "14min 56s ± 52.7 s per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
+     ]
+    },
+    {
+     "data": {
+      "text/plain": [
+       "<TimeitResult : 14min 56s ± 52.7 s per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
+      ]
+     },
+     "execution_count": 12,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
+   "source": [
+    "%%timeit -n 1 -r 3 -o\n",
+    "\n",
+    "queries = _generate_singling_out_queries(\n",
+    "    df=evaluator._syn,\n",
+    "    n_attacks=evaluator._n_attacks,\n",
+    "    n_cols=evaluator._n_cols,\n",
+    "    mode=mode,\n",
+    "    max_attempts=evaluator._max_attempts,\n",
+    ")\n",
+    "_queries = _evaluate_queries(df=evaluator._ori, queries=queries)"
+   ]
+  }
+ ],
+ "metadata": {
+  "kernelspec": {
+   "display_name": "Python 3 (ipykernel)",
+   "language": "python",
+   "name": "python3"
+  },
+  "language_info": {
+   "codemirror_mode": {
+    "name": "ipython",
+    "version": 3
+   },
+   "file_extension": ".py",
+   "mimetype": "text/x-python",
+   "name": "python",
+   "nbconvert_exporter": "python",
+   "pygments_lexer": "ipython3",
+   "version": "3.10.16"
+  }
+ },
+ "nbformat": 4,
+ "nbformat_minor": 2
+}