--- a/notebooks/measure_time.ipynb
+++ b/notebooks/measure_time.ipynb
@@ -9,7 +9,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 13,
+   "execution_count": 1,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -31,7 +31,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 2,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -51,15 +51,16 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 10,
    "metadata": {},
    "outputs": [],
    "source": [
-    "max_attempts = 1_000_000\n",
+    "max_attempts = 100_000\n",
     "evaluator = SinglingOutEvaluator(ori=ori, \n",
     "                                 syn=syn, \n",
     "                                 control=control,\n",
     "                                 n_attacks=max_attempts,\n",
+    "                                 n_cols=12,\n",
     "                                 max_attempts=max_attempts)"
    ]
   },
@@ -72,7 +73,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 4,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -81,23 +82,35 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 12,
+   "execution_count": 11,
    "metadata": {},
    "outputs": [
+    {
+     "name": "stderr",
+     "output_type": "stream",
+     "text": [
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 70238 instead of the requested 100000.\n",
+      "Attack `multivariate` could generate only 70238 singling out queries out of the requested 100000. This can probably lead to an underestimate of the singling out risk.\n",
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 70002 instead of the requested 100000.\n",
+      "Attack `multivariate` could generate only 70002 singling out queries out of the requested 100000. This can probably lead to an underestimate of the singling out risk.\n",
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 69771 instead of the requested 100000.\n",
+      "Attack `multivariate` could generate only 69771 singling out queries out of the requested 100000. This can probably lead to an underestimate of the singling out risk.\n"
+     ]
+    },
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "14min 56s ± 52.7 s per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
+      "1min 44s ± 2.43 s per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<TimeitResult : 14min 56s ± 52.7 s per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
+       "<TimeitResult : 1min 44s ± 2.43 s per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
       ]
      },
-     "execution_count": 12,
+     "execution_count": 11,
      "metadata": {},
      "output_type": "execute_result"
     }