--- a/notebooks/measure_time.ipynb
+++ b/notebooks/measure_time.ipynb
@@ -15,8 +15,6 @@
    "source": [
     "import os\n",
     "import pandas as pd\n",
-    "import matplotlib.pyplot as plt\n",
-    "import seaborn as sns\n",
     "\n",
     "from anonymeter.evaluators import SinglingOutEvaluator\n",
     "from anonymeter.evaluators.singling_out_evaluator import _evaluate_queries, _generate_singling_out_queries"
@@ -51,17 +49,15 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 10,
+   "execution_count": 3,
    "metadata": {},
    "outputs": [],
    "source": [
-    "max_attempts = 100_000\n",
     "evaluator = SinglingOutEvaluator(ori=ori, \n",
     "                                 syn=syn, \n",
     "                                 control=control,\n",
-    "                                 n_attacks=max_attempts,\n",
-    "                                 n_cols=12,\n",
-    "                                 max_attempts=max_attempts)"
+    "                                 n_attacks=2000,\n",
+    "                                 n_cols=12)"
    ]
   },
   {
@@ -82,35 +78,23 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 11,
+   "execution_count": 5,
    "metadata": {},
    "outputs": [
-    {
-     "name": "stderr",
-     "output_type": "stream",
-     "text": [
-      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 70238 instead of the requested 100000.\n",
-      "Attack `multivariate` could generate only 70238 singling out queries out of the requested 100000. This can probably lead to an underestimate of the singling out risk.\n",
-      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 70002 instead of the requested 100000.\n",
-      "Attack `multivariate` could generate only 70002 singling out queries out of the requested 100000. This can probably lead to an underestimate of the singling out risk.\n",
-      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 69771 instead of the requested 100000.\n",
-      "Attack `multivariate` could generate only 69771 singling out queries out of the requested 100000. This can probably lead to an underestimate of the singling out risk.\n"
-     ]
-    },
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "1min 44s ± 2.43 s per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
+      "3.01 s ± 105 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<TimeitResult : 1min 44s ± 2.43 s per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
+       "<TimeitResult : 3.01 s ± 105 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
       ]
      },
-     "execution_count": 11,
+     "execution_count": 5,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -126,6 +110,113 @@
     "    max_attempts=evaluator._max_attempts,\n",
     ")\n",
     "_queries = _evaluate_queries(df=evaluator._ori, queries=queries)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 6,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "mean  : 3.01 s\n",
+      "std   : 0.10 s\n"
+     ]
+    }
+   ],
+   "source": [
+    "print(f\"mean  : {_.average:.2f} s\")\n",
+    "print(f\"std   : {_.stdev:.2f} s\")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Repeat the above with n_cols=3"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 7,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "evaluator = SinglingOutEvaluator(ori=ori, \n",
+    "                                 syn=syn, \n",
+    "                                 control=control,\n",
+    "                                 n_attacks=2000,\n",
+    "                                 max_attempts=100_000,\n",
+    "                                 n_cols=3)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 8,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stderr",
+     "output_type": "stream",
+     "text": [
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 314 instead of the requested 2000.\n",
+      "Attack `multivariate` could generate only 314 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n",
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 277 instead of the requested 2000.\n",
+      "Attack `multivariate` could generate only 277 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n",
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 296 instead of the requested 2000.\n",
+      "Attack `multivariate` could generate only 296 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n"
+     ]
+    },
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "17.5 s ± 26.8 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
+     ]
+    },
+    {
+     "data": {
+      "text/plain": [
+       "<TimeitResult : 17.5 s ± 26.8 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
+      ]
+     },
+     "execution_count": 8,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
+   "source": [
+    "%%timeit -n 1 -r 3 -o\n",
+    "\n",
+    "queries = _generate_singling_out_queries(\n",
+    "    df=evaluator._syn,\n",
+    "    n_attacks=evaluator._n_attacks,\n",
+    "    n_cols=evaluator._n_cols,\n",
+    "    mode=mode,\n",
+    "    max_attempts=evaluator._max_attempts,\n",
+    ")\n",
+    "_queries = _evaluate_queries(df=evaluator._ori, queries=queries)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 9,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "mean  : 17.46 s\n",
+      "std   : 0.03 s\n"
+     ]
+    }
+   ],
+   "source": [
+    "print(f\"mean  : {_.average:.2f} s\")\n",
+    "print(f\"std   : {_.stdev:.2f} s\")"
    ]
   }
  ],