--- a/notebooks/measure_time_singling_out.ipynb
+++ b/notebooks/measure_time_singling_out.ipynb
@@ -9,7 +9,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 1,
+   "execution_count": 14,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -29,7 +29,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 15,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -49,7 +49,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 16,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -69,7 +69,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 4,
+   "execution_count": 17,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -78,23 +78,23 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 18,
    "metadata": {},
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "3.01 s ± 105 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
+      "2.86 s ± 76.3 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<TimeitResult : 3.01 s ± 105 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
+       "<TimeitResult : 2.86 s ± 76.3 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
       ]
      },
-     "execution_count": 5,
+     "execution_count": 18,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -108,21 +108,22 @@
     "    n_cols=evaluator._n_cols,\n",
     "    mode=mode,\n",
     "    max_attempts=evaluator._max_attempts,\n",
+    "    rng=evaluator._rng\n",
     ")\n",
     "_queries = _evaluate_queries(df=evaluator._ori, queries=queries)"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 19,
    "metadata": {},
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "mean  : 3.01 s\n",
-      "std   : 0.10 s\n"
+      "mean  : 2.86 s\n",
+      "std   : 0.08 s\n"
      ]
     }
    ],
@@ -140,7 +141,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": 20,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -154,35 +155,35 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 8,
+   "execution_count": 21,
    "metadata": {},
    "outputs": [
     {
      "name": "stderr",
      "output_type": "stream",
      "text": [
-      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 314 instead of the requested 2000.\n",
-      "Attack `multivariate` could generate only 314 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n",
-      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 277 instead of the requested 2000.\n",
-      "Attack `multivariate` could generate only 277 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n",
-      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 296 instead of the requested 2000.\n",
-      "Attack `multivariate` could generate only 296 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n"
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 322 instead of the requested 2000.\n",
+      "Attack `multivariate` could generate only 322 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n",
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 278 instead of the requested 2000.\n",
+      "Attack `multivariate` could generate only 278 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n",
+      "Reached maximum number of attempts 100000 when generating singling out queries. Returning 312 instead of the requested 2000.\n",
+      "Attack `multivariate` could generate only 312 singling out queries out of the requested 2000. This can probably lead to an underestimate of the singling out risk.\n"
      ]
     },
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "17.5 s ± 26.8 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
+      "17.5 s ± 411 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)\n"
      ]
     },
     {
      "data": {
       "text/plain": [
-       "<TimeitResult : 17.5 s ± 26.8 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
+       "<TimeitResult : 17.5 s ± 411 ms per loop (mean ± std. dev. of 3 runs, 1 loop each)>"
       ]
      },
-     "execution_count": 8,
+     "execution_count": 21,
      "metadata": {},
      "output_type": "execute_result"
     }
@@ -196,21 +197,22 @@
     "    n_cols=evaluator._n_cols,\n",
     "    mode=mode,\n",
     "    max_attempts=evaluator._max_attempts,\n",
+    "    rng=evaluator._rng\n",
     ")\n",
     "_queries = _evaluate_queries(df=evaluator._ori, queries=queries)"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 9,
+   "execution_count": 22,
    "metadata": {},
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "mean  : 17.46 s\n",
-      "std   : 0.03 s\n"
+      "mean  : 17.50 s\n",
+      "std   : 0.41 s\n"
      ]
     }
    ],