… notebooks/singling_out.ipynb