The singling out attack implementation is slow because queries are constructed and evaluated sequentially, one-by-one. 

In this pull request, queries are instead evaluated in batches, in parallel. This is enabled by [polars](https://pola.rs/), an efficient pandas alternative. The SinglingOutEvaluator still accepts pandas dataframes but they are converted to polars dataframes that are used to construct and evaluate queries.

Time was measured with the code in ``notebooks/measure_time_singling_out.ipynb``: 2000 queries are constructed for the Adult dataset using either 3 or 12 auxiliary columns (the minimum and the maximum number of auxiliary columns used in the practical attack in the Anonymeter paper). Query generation and evaluation is repeated 3 times, mean and standard deviation are reported both for the current pandas implementation and the proposed polars implementation. The Standard_E4d_v4 Azure machine (4 CPU cores) was used (using a CPU with more cores is expected to lead to even larger speed-ups).

The results:

| Number of columns | Pandas (mean ± std) | Polars (mean ± std) | Speed-up |
|-------------------|---------------------|---------------------|----------------------------|
| 3                 | 527.89 ± 1.34 s     | 17.46 ± 0.03 s      | 30.2×                     |
| 12                | 61.20 ± 0.03 s      | 3.01 ± 0.10 s       | 20.3×                     |

Additionally, ``seed`` parameter has been added to the SinglingOutEvaluator to enable reproduciblity, and a bug has been fixed where in the case that fewer queries were successfully constructed than expected, the expected rather than actual number of queries was used in the risk estimation, which led to underestimation.

To ensure that the code functions correctly (in addition to tests passing), the old and the new implementations were used to estimate the risks of the synthetic data included in the repository, and artificially constructed synthetic data with 20% leakage. The code in ``notebooks/singling_out.ipynb`` was used, with each experiment repeated 5 times. The results are similar between the implementations.

|Data | Pandas | Polars |
|---|---|---|
|CTGAN| ![image](https://github.com/user-attachments/assets/f565303f-703b-4a40-9944-e69c25747813) | ![image](https://github.com/user-attachments/assets/a545e54f-37ab-4c84-b3d6-c53c02c5dec4) |
|Artificial| ![image](https://github.com/user-attachments/assets/ad949632-5da5-4825-b3c2-c6b07187eea1) | ![image](https://github.com/user-attachments/assets/387e883c-0608-40a5-acb7-76e384b122af) |