@AwesomeLemon 2025-06-18
Thank you for taking the time to give feedback on my code! 

I've implemented most of it, here are the key changes:

1. UniqueSinglingOutQueries modified to evaluate queries inside.
2. Fixed an issue where an all-null column would throw an error in univariate query construction due to min/max calls.
3. Random generator is specified in one place and passed to the functions that need it.

I've resolved the comments with your suggestions that I've straightforwardly implemented. I've answered to the comments where the changes I made were not completely trivial. 

The only strong point of disagreement is about dividing by the required/actual number of attacks - I've jotted down my thoughts as a response to your comment there.

I'm a bit confused why you can't see the notebooks: I see them both in this pull request, and in my fork (see [singling_out.py](https://github.com/AwesomeLemon/anonymeter/blob/7b6e6937a5e7e8a6c7127887ba702c080730e0eb/notebooks/singling_out.ipynb) and [measure_time_singling_out.ipynb](https://github.com/AwesomeLemon/anonymeter/blob/7b6e6937a5e7e8a6c7127887ba702c080730e0eb/notebooks/measure_time_singling_out.ipynb))