@MatteoGiomi 2025-06-19
Hi @AwesomeLemon, thanks again for your code. While sleeping over it, two questions came to my mind: 
- how (if possible) can the user control the amount of paralelization introduced by `polars`?
-