@MatteoGiomi 2025-06-19
one question about `polars`. If it possible (and if so how) to control the amount of parallelization? I might have missed it in the code..