@AwesomeLemon 2025-06-19
> one question about `polars`. Is it possible (and if so how) to control the amount of parallelization? I might have missed it in the code..

By default, polars uses all the cores. To reduce the number of used cores, it seems that the ``POLARS_MAX_THREADS`` environment variable has to be exported (https://github.com/pola-rs/polars/issues/1718)